% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{add_predictions}
\alias{add_predictions}
\title{Add predictions to a data frame}
\usage{
add_predictions(data, ...)
}
\arguments{
\item{data}{A data frame used to generate the predictions.}

\item{...}{Name-value pairs giving the variable name and model.}
}
\description{
Add predictions to a data frame
}
\examples{
m1 <- lm(mpg ~ wt, data = mtcars)
m2 <- lm(mpg ~ poly(wt, 2), data = mtcars)
mtcars \%>\%
  add_predictions(m1 = m1, m2 = m2)
}

