% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{add_residuals}
\alias{add_residuals}
\title{Add residuals to a data frame}
\usage{
add_residuals(data, ...)
}
\arguments{
\item{data}{A data frame used to generate the predictions.}

\item{...}{Name-value pairs giving the variable name and model.}
}
\description{
Add residuals to a data frame
}
\examples{
m1 <- lm(mpg ~ wt, data = mtcars)
m2 <- lm(mpg ~ poly(wt, 2), data = mtcars)
mtcars \%>\%
  add_residuals(m1 = m1, m2 = m2)
}

