% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Generate a boostrap replicate}
\usage{
bootstrap(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Generate a boostrap replicate
}
\examples{
coef(lm(mpg ~ wt, data = bootstrap(mtcars)))

library(purrr)
bootstraps <- rerun(100, bootstrap(mtcars))
models <- map(bootstraps, ~ lm(mpg ~ wt, data = .))
tidied <- map_df(models, broom::tidy, .id = "id")

hist(subset(tidied, term == "wt")$estimate)
hist(subset(tidied, term == "(Intercept)")$estimate)
}

