% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ops.R
\name{deepcopy}
\alias{deepcopy}
\title{Deep Copy a ModelObject}
\usage{
deepcopy(mo)
}
\arguments{
\item{mo}{a ModelObject}
}
\value{
A copy of the ModelObject \code{mo} that *should* be a separate object from
the original
}
\description{
Deep Copy a ModelObject
}

