% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-model.R
\name{expand_model}
\alias{expand_model}
\title{Expand a grid for a model}
\usage{
expand_model(data, model, ...)
}
\arguments{
\item{data}{A data frame}

\item{model}{A model}

\item{...}{Variables passed on to \code{\link[tidyr]{expand}()}}
}
\description{
This is a wrapper around \code{\link[tidyr]{expand}()} that automatically
fills in variables not otherwise provided with their typical values.
}
\examples{
mod <- lm(mpg ~ wt + cyl + vs, data = mtcars)
expand_model(mtcars, mod)
expand_model(mtcars, mod, cyl = seq_range(cyl, 9))
}

