% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ref_line.R
\name{geom_ref_line}
\alias{geom_ref_line}
\title{Add a reference line (ggplot2).}
\usage{
geom_ref_line(h, v, size = 2, colour = "white")
}
\arguments{
\item{h, v}{Position of horizontal or vertical reference line}

\item{size}{Line size}

\item{colour}{Line colour}
}
\description{
Add a reference line (ggplot2).
}

