% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{mcmc_cv}
\alias{mcmc_cv}
\title{Generate \code{n} MCMC cross-validated test-training pairs}
\usage{
mcmc_cv(data, n, test = 0.1)
}
\arguments{
\item{data}{A data frame}

\item{n}{Number of test-training pairs to generate}

\item{test}{Proportion of observations that should be held out for testing.}
}
\value{
A data frame with \code{n} rows and columns \code{test} and
  \code{train}. These are list columns containing \code{\link{resample}}
  objects.
}
\description{
Generate \code{n} MCMC cross-validated test-training pairs
}
\examples{
cv <- mcmc_cv(mtcars, 100)

library(purrr)
models <- map(cv$train, ~ lm(mpg ~ wt, data = .))
errs <- map2_dbl(models, cv$test, rmse)
hist(errs)
}

