% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{mod_lm}
\alias{mod_lm}
\title{Ordinary Least Squares}
\usage{
mod_lm(data, formula, robust = FALSE)
}
\arguments{
\item{data}{A data.frame containing the dependent and independent variables used in the regression.}

\item{formula}{an object of class \code{"\link{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{robust}{Logical. If true robust standard errors is used.}
}
\value{
mod_lm returns an object of the class mod_lm, which is a list containing information about the fitted model.
}
\description{
Ordinary Least Squares
}
\examples{
mod_lm(mtcars, mpg~cyl)
}
