% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{model-quality}
\alias{model-quality}
\alias{mse}
\alias{rmse}
\alias{mae}
\alias{rsquare}
\alias{qae}
\title{Compute model quality for a given dataset}
\usage{
mse(model, data)

rmse(model, data)

mae(model, data)

rsquare(model, data)

qae(model, data, probs = c(0.05, 0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{model}{A model}

\item{data}{The dataset}

\item{probs}{Numeric vector of probabilities}
}
\description{
Three summaries can be interpreted on the scale of the response:
\itemize{
\item \code{rmse()} is the root-mean-squared-error
\item \code{mae()} is the mean absolute error
\item \code{qae()} is quantiles of absolute error.
}

Other summaries have varying scales:
\itemize{
\item \code{mse()} is the mean-squared-error and can be interpreted on the squared
scale of the response.
\item \code{rsquare()} is the variance of the predictions divided by the
variance of the response.
}
}
\examples{
mod <- lm(mpg ~ wt, data = mtcars)
mse(mod, mtcars)
rmse(mod, mtcars)
rsquare(mod, mtcars)
mae(mod, mtcars)
qae(mod, mtcars)
}
