% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelObject.R
\docType{class}
\name{ModelObject-class}
\alias{ModelObject}
\alias{ModelObject-class}
\title{ModelObject RefClass Object}
\description{
ModelObject RefClass Object
}
\section{Fields}{

\describe{
\item{\code{name}}{Character string}

\item{\code{expr}}{R expression that defines the model equation}

\item{\code{P}}{a list of named numeric vectors defining parameters:
     \code{list(p0=<initial guess>, lb=<lower bounds>, ub=<upper bounds>)}}
}}
\section{Methods}{

\describe{
\item{\code{initialize(...)}}{initialize object}

\item{\code{jacobian(p = NULL, data)}}{Evaluates the model jacobian}

\item{\code{setP(p0 = NULL, lb = NULL, ub = NULL)}}{Sets initial parameter values, and optionally upper and lower bounds.
Arguments p0, lb, and ub should be named numeric vectors}

\item{\code{value(p = NULL, data)}}{Evaluates the model}
}}

