% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelObjectUtils.R
\name{ModelObjectFit}
\alias{ModelObjectFit}
\title{Generalized non-linear model fitting}
\usage{
ModelObjectFit(data, MODEL, weightFun = NULL, optimizer = c("nlminb",
  "trust", "hybrid"), control = NULL, trust.bounded = F)
}
\arguments{
\item{data}{input data.frame or list to fit}

\item{MODEL}{an object of class \link{ModelObject} defining the model to fit}

\item{optimizer}{Name of the optimizer to use for minimization. Default 'nlminb'.
Optimizer 'trust' requires installing the 'trust' package.}

\item{control}{control list to pass to \link{nlminb}. Ignored if using 'trust'.}
}
\value{
A \code{fit} list-object as returned by \link{nlminb} with additional elements:
\describe{
   \item{residuals}{A numeric vector of residuals}
   \item{DATA}{
     A copy of the data used to fit model parameters
   }
   \item{MODEL}{
     A copy of \code{MODEL} provided to the optimization.
     Used for evaulating the model prediction using the fitted parameters
   }
}
}
\description{
Fits a non-linear model using minimization of squared residuals
}
\details{
MODEL is expected to be of the form:
\deqn{y = f(p, x)}

Thus, data to be fitted must have a named element \code{x} indicating the
independent variable and a named element \code{y} indicating the dependent
(observed) variable.
}
\seealso{
\link{nlminb}
}

