% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelObjectUtils.R
\name{ModelObjectPredict}
\alias{ModelObjectPredict}
\title{Generic model fit predictor}
\usage{
ModelObjectPredict(fit, data = NULL)
}
\arguments{
\item{fit}{A \code{fit} object as returned by \code{ModelFit} which is expected
   to have an element of \code{MODEL$modelFun}}
}
\value{
An \code{xy.coords} compatible list object of the fitted model
}
\description{
Predicts a fitted model over the x-values provided in the data.frame or list \code{data}
}

