% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelR.R
\docType{package}
\name{modelR-package}
\alias{modelR}
\alias{modelR-package}
\title{modelR: Statistical modelling made easy}
\description{
Statistical modelling made easy
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elben10/modelR}
  \item Report bugs at \url{https://github.com/elben10/modelR/issues}
}

}
\author{
\strong{Maintainer}: Jakob Jul Elben \email{elbenjakobjul@gmail.com}

}
