% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelr.R
\docType{package}
\name{modelr}
\alias{modelr}
\alias{modelr-package}
\title{Object oriented model based optimization}
\description{
Provides classes for model object construction that automates
calculation of analytical gradients, jacobians, and hessians for faster
optimization.
}
\seealso{
\link{stats}{nlminb}, \link{trust:trust-package}
}

