% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition a dataset into test and training splits}
\usage{
partition(data, p)
}
\arguments{
\item{data}{A data frame}

\item{p}{A named numeric vector giving where the value is the probability
that an observation will be assigned to that group.}
}
\description{
Partition a dataset into test and training splits
}
\examples{
ex <- partition(mtcars, c(test = 0.3, train = 0.7))
mod <- lm(mpg ~ wt, data = ex$train)
rmse(mod, ex$test)
rmse(mod, ex$train)

# Repeat 100 times
library(purrr)
crossv <- transpose(rerun(100, partition(mtcars, c(test = 0.3, train = 0.7))))
models <- map(crossv$train, ~ lm(mpg ~ wt, data = .))
errs <- map2_dbl(models, crossv$test, rmse)
hist(errs)
}

