% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{resample_bootstrap}
\alias{resample_bootstrap}
\title{Generate a boostrap replicate}
\usage{
resample_bootstrap(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Generate a boostrap replicate
}
\examples{
coef(lm(mpg ~ wt, data = resample_bootstrap(mtcars)))
coef(lm(mpg ~ wt, data = resample_bootstrap(mtcars)))
coef(lm(mpg ~ wt, data = resample_bootstrap(mtcars)))
}
\seealso{
Other resampling techniques: 
\code{\link{bootstrap}()},
\code{\link{resample_partition}()},
\code{\link{resample}()}
}
\concept{resampling techniques}
