% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_range.R
\name{seq_range}
\alias{seq_range}
\title{Generate a sequence over the range of a vector}
\usage{
seq_range(x, n, trim = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{Length of sequence to generate}

\item{trim}{Optionally, trim values off the tails.
\code{trim / 2 * length(x)} values are removed from each tail.}
}
\description{
Generate a sequence over the range of a vector
}
\examples{
x <- rcauchy(100)
seq_range(x, 10)
seq_range(x, 10, trim = 0.1)
}

