% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ops.R
\name{wrap}
\alias{wrap}
\title{Wrap ModelObjects with transformation functions (NSE)}
\usage{
wrap(..., fun = identity)
}
\arguments{
\item{...}{list of ModelObjects}

\item{fun}{function with which models are wrapped. Default is \code{identity}}
}
\value{
A list of ModelObjects with names and expressions wrapped with the function specified in \code{fun}.
}
\description{
Wrap ModelObjects with transformation functions (NSE)
}

