% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelr.R
\docType{package}
\name{modelr-package}
\alias{modelr}
\alias{modelr-package}
\title{modelr: Modelling Functions that Work with the Pipe}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions for modelling that help you seamlessly integrate modelling into a pipeline of data manipulation and visualisation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://modelr.tidyverse.org}
  \item \url{https://github.com/tidyverse/modelr}
  \item Report bugs at \url{https://github.com/tidyverse/modelr/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
