% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\alias{sim1}
\alias{sim2}
\alias{sim3}
\alias{sim4}
\title{Simple simulated datasets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 2 columns.}
\usage{
sim1

sim2

sim3

sim4
}
\description{
These simple simulated datasets are useful for teaching modelling
basics.
}
\keyword{datasets}
